refactored out the initial list of components and imports that are not used
