+++ src/constants/Api.js
@@ -1,5 +1,4 @@
 export const apiBaseUrl = 'https://min-api.cryptocompare.com/data/price?fsym=';
 export const apiCurrencyResponseUrl = '&tsyms=USD,CAD,ETH,BTC,EUR';
-
 export const apiMultipleCurrencyBaseUrl = 'https://min-api.cryptocompare.com/data/pricemulti?fsyms=';
 export const apiMulitpleResponseUrl = '&tsyms=USD,CAD,ETH,BTC,EUR';
