+++ src/constants/data/Terms.js
